Move QuickSearch to Shared Package
