+++ frontend/packages/console-shared/locales/en/console-shared.json
@@ -134,6 +134,11 @@
   "Increase the Pod count": "Increase the Pod count",
   "Decrease the Pod count": "Decrease the Pod count",
   "Close": "Close",
+  "Quick search bar": "Quick search bar",
+  "No results": "No results",
+  "Provided by {{provider}}": "Provided by {{provider}}",
+  "Quick search list": "Quick search list",
+  "Quick search": "Quick search",
   "Hover": "Hover",
   "Drag": "Drag",
   "Click": "Click",
