+++ frontend/packages/console-shared/src/components/index.ts
@@ -28,3 +28,4 @@ export * from './custom-resource-list';
 export * from './toast';
 export * from './markdown-extensions';
 export * from './actions';
+export * from './quick-search';
