+++ frontend/packages/console-shared/src/components/quick-search/QuickSearchBar.scss
