+++ frontend/packages/console-shared/src/components/quick-search/QuickSearchContent.tsx
