+++ frontend/packages/console-shared/src/components/quick-search/QuickSearchDetails.scss
