+++ frontend/packages/console-shared/src/components/quick-search/QuickSearchIcon.tsx
