+++ frontend/packages/console-shared/src/components/quick-search/QuickSearchList.scss
