+++ frontend/packages/console-shared/src/components/quick-search/QuickSearchModalBody.scss
