+++ frontend/packages/console-shared/src/components/quick-search/index.ts
@@ -1,0 +1,3 @@
+export { default as QuickSearchController } from './QuickSearchController';
+export * from './utils/quick-search-types';
+export * from './utils/quick-search-utils';
