+++ frontend/packages/console-shared/src/components/quick-search/utils/quick-search-types.ts
