+++ frontend/packages/topology/locales/en/topology.json
@@ -60,19 +60,14 @@
   "Shortcuts": "Shortcuts",
   "View shortcuts": "View shortcuts",
   "Drop file ({{fileTypes}}) here": "Drop file ({{fileTypes}}) here",
-  "View all developer catalog items ({{itemCount, number}})": "View all developer catalog items ({{itemCount, number}})",
-  "View all quick starts ({{itemCount, number}})": "View all quick starts ({{itemCount, number}})",
-  "View all samples ({{itemCount, number}})": "View all samples ({{itemCount, number}})",
-  "Quick search bar": "Quick search bar",
-  "Add to Project": "Add to Project",
-  "No results": "No results",
-  "Quick search button": "Quick search button",
-  "Provided by {{provider}}": "Provided by {{provider}}",
-  "Quick search list": "Quick search list",
-  "Quick search": "Quick search",
   "Prerequisites": "Prerequisites",
   "Quick Starts": "Quick Starts",
   "Start": "Start",
+  "View all developer catalog items ({{itemCount, number}})": "View all developer catalog items ({{itemCount, number}})",
+  "View all quick starts ({{itemCount, number}})": "View all quick starts ({{itemCount, number}})",
+  "View all samples ({{itemCount, number}})": "View all samples ({{itemCount, number}})",
+  "Add to Project": "Add to Project",
+  "Quick search button": "Quick search button",
   "Visual connector": "Visual connector",
   "Binding connector": "Binding connector",
   "Traffic connector": "Traffic connector",
