--- frontend/packages/topology/src/components/quick-search/QuickSearchButton.scss
@@ -1,3 +1,0 @@
-.odc-quick-search-button {
-  padding: 0px;
-}
