+++ frontend/packages/topology/src/components/quick-search/TopologyQuickSearchButton.scss
@@ -1,0 +1,3 @@
+.odc-topology-quick-search-button {
+  padding: 0px;
+}
